requirements.txt
wheel
nbresult
colorama
ipdb
ipykernel
yapf
matplotlib
pygeohash
pytest
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1
google-auth
db-dtypes
pyarrow
bokeh==2.4.2
distributed==2021.5.0
dateparser==1.1.1
dask[complete]==2021.5.0
fastparquet==0.4.1
httpcore==1.0.4
nlpaug==1.0.1

# prevent bq 3+ db-dtypes package error when running training from bq
ipython

python-dotenv
prefect==2.8.1
psycopg2-binary

# API


# jupiter notebook
notebook
ipython
torch
ipywidgets


# Dependancies for sheethelper
gspread
transformers
openai
wordcloud
streamlit==1.32.0
sentencepiece
requests
black
skimpy
streamlit-shadcn-ui
loguru
schedule
wordcloud
requests
nlpretext
plotly

